+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+## [1.2.0] - 2026-08-12
+
+### Added
+
 ## [1.1.0] - 2026-08-12
 
 ### Added